Add testing config and lint errors (#10)
* add tests: jest config and common testing
* remove unnecessary file
* lint: renamed files, apply ESLint
* group similar features in same file
