+++ src/plugins/profiling/common/index.test.ts
@@ -1,0 +1,21 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import { timeRangeFromRequest } from './index';
+
+describe('Common profiling helpers', () => {
+  test('convert query parameters time ranges into tuple', () => {
+    const request = {
+      query: {
+        timeFrom: 123,
+        timeTo: 456,
+      },
+    };
+    expect(timeRangeFromRequest(request)).toEqual([123, 456]);
+  });
+});
