+++ src/plugins/profiling/server/routes/load_flamechart.ts
@@ -6,15 +6,16 @@
  * Side Public License, v 1.
  */
 import { schema } from '@kbn/config-schema';
+import type { IRouter } from 'kibana/server';
 import type { DataRequestHandlerContext } from '../../../data/server';
-import type { IRouter } from '../../../../core/server';
 import { getLocalRoutePaths, timeRangeFromRequest } from '../../common';
+import { mapFlamechart } from './mappings';
 
-export function registerTraceEventsTopNThreadsRoute(router: IRouter<DataRequestHandlerContext>) {
+export function registerFlameChartElasticRoute(router: IRouter<DataRequestHandlerContext>) {
   const paths = getLocalRoutePaths();
   router.get(
     {
-      path: paths.TopNThreads,
+      path: paths.FlamechartElastic,
       validate: {
         query: schema.object({
           index: schema.maybe(schema.string()),
@@ -26,7 +27,30 @@ export function registerTraceEventsTopNThreadsRoute(router: IRouter<DataRequestH
     },
     async (ctx, request, response) => {
       const [timeFrom, timeTo] = timeRangeFromRequest(request);
-      const src = await import(`../fixtures/threads_${timeTo - timeFrom}`);
+      const src = await import(`../fixtures/flamechart_${timeTo - timeFrom}`);
+      delete src.default;
+      return response.ok({ body: mapFlamechart(src) });
+    }
+  );
+}
+
+export function registerFlameChartPixiRoute(router: IRouter<DataRequestHandlerContext>) {
+  const paths = getLocalRoutePaths();
+  router.get(
+    {
+      path: paths.FlamechartPixi,
+      validate: {
+        query: schema.object({
+          index: schema.maybe(schema.string()),
+          projectID: schema.maybe(schema.string()),
+          timeFrom: schema.maybe(schema.string()),
+          timeTo: schema.maybe(schema.string()),
+        }),
+      },
+    },
+    async (ctx, request, response) => {
+      const [timeFrom, timeTo] = timeRangeFromRequest(request);
+      const src = await import(`../fixtures/flamechart_${timeTo - timeFrom}`);
       delete src.default;
       return response.ok({ body: src });
     }
